refactor: trying to get the ctype utils to work
